veasypooler 안쓰게 변경
instantiate는 바로 스크립트 가져올 수 있는데, 풀러쓰면 못해서 Get Component 해야함

shootable editor 추가

+++ Assets/Resources/Script/Operate/Shootable/Shootable.cs
@@ -8,14 +8,16 @@ public class Shootable : Operable
     [System.NonSerialized]
     public Unit target;
 
-    // TODO 에디터 이용해서 userange 켰을때만 range 보이게
     public bool isRangeless;
     public float range;
 
-    public float damage;
+    public int damage;
 
     public float attackDelay;
     private float remainAttackDelay;
+
+    // 활성/비활성
+    // 비활성 중 재사용 대기시간 감소 여부
 
     void Start()
     {
